moved single functions into agile file
